+++ e2e/reporters/recorder.js
@@ -89,11 +89,6 @@ class E2eRecorderReporter extends JestMetadataReporter {
       await debugUtils.aggregateLogs();
     }
   }
-
-  static onTestEnvironmentCreate(env) {
-    console.log('I AM ALIVE');
-    console.log('  ..--..');
-  }
 }
 
 function sleep(ms) {
