+++ environment-decorator.js
@@ -1,2 +1,2 @@
 /* Jest 27 fallback */
-module.exports = require('./dist/environment-listener');
+module.exports = require('./dist/environment-decorator');
