+++ environment-listener.js
@@ -1,0 +1,2 @@
+/* Jest 27 fallback */
+module.exports = require('./dist/environment-listener');
