+++ src/utils/emitters/SerialEmitter.test.ts
@@ -1,8 +1,8 @@
-import { SerialSyncEmitter } from './SerialSyncEmitter';
+import { SerialEmitter } from './SerialEmitter';
 
-describe('SerialSyncEmitter', () => {
+describe('SerialEmitter', () => {
   it('should emit events', () => {
-    const emitter = new SerialSyncEmitter<TestEvent>();
+    const emitter = new SerialEmitter<TestEvent>();
     const listener = jest.fn();
     emitter.on('test', listener);
 
@@ -16,7 +16,7 @@ describe('SerialSyncEmitter', () => {
   });
 
   it('should allow subscribing to events only once', () => {
-    const emitter = new SerialSyncEmitter<TestEvent>();
+    const emitter = new SerialEmitter<TestEvent>();
     const listener = jest.fn();
     emitter.once('test', listener);
     emitter.emit({ type: 'test', payload: 42 });
@@ -26,7 +26,7 @@ describe('SerialSyncEmitter', () => {
   });
 
   it('should allow subscribing to all events', () => {
-    const emitter = new SerialSyncEmitter<TestEvent>();
+    const emitter = new SerialEmitter<TestEvent>();
     const listener = jest.fn();
     emitter.on('*', listener);
     emitter.emit({ type: 'test', payload: 42 });
@@ -38,7 +38,7 @@ describe('SerialSyncEmitter', () => {
   });
 
   it('should allow unsubscribing from events', () => {
-    const emitter = new SerialSyncEmitter<TestEvent>();
+    const emitter = new SerialEmitter<TestEvent>();
     const listener = jest.fn();
     emitter.on('test', listener);
     emitter.emit({ type: 'test', payload: 42 });
@@ -49,7 +49,7 @@ describe('SerialSyncEmitter', () => {
   });
 
   it('should delay emits within emits', () => {
-    const emitter = new SerialSyncEmitter<TestEvent>();
+    const emitter = new SerialEmitter<TestEvent>();
     const listener1 = jest.fn(() => emitter.emit({ type: 'test', payload: 84 }));
     const listener2 = jest.fn();
     emitter.once('test', listener1);
