+++ src/utils/emitters/index.ts
@@ -1,4 +1,2 @@
 export * from './AggregatedEmitter';
-export * from './SerialAsyncEmitter';
-export * from './SerialSyncEmitter';
-export * from './SemiAsyncEmitter';
+export * from './SerialEmitter';
