+++ src/utils/logger.ts
@@ -8,16 +8,18 @@ import { noop } from './noop';
 const logsDirectory = process.env.JEST_METADATA_DEBUG;
 
 bunyamin.logger = createBunyanImpl(isTraceEnabled());
-bunyamin.threadGroups.push(
-  { id: 'ipc-server', displayName: 'IPC Server (jest-metadata)' },
-  { id: 'ipc-client', displayName: 'IPC Client (jest-metadata)' },
-  { id: 'emitter-core', displayName: 'Core emitter (jest-metadata)' },
-  { id: 'emitter-set', displayName: 'Set emitter (jest-metadata)' },
-  { id: 'emitter-events', displayName: 'Events emitter (jest-metadata)' },
-  { id: 'environment', displayName: 'Test Environment (jest-metadata)' },
-  { id: 'metadata', displayName: 'Metadata (jest-metadata)' },
-  { id: 'reporter', displayName: 'Reporter (jest-metadata)' },
-);
+if (!bunyamin.threadGroups.some((x) => x.id === 'metadata')) {
+  bunyamin.threadGroups.push(
+    { id: 'ipc-server', displayName: 'IPC Server (jest-metadata)' },
+    { id: 'ipc-client', displayName: 'IPC Client (jest-metadata)' },
+    { id: 'emitter-core', displayName: 'Core emitter (jest-metadata)' },
+    { id: 'emitter-set', displayName: 'Set emitter (jest-metadata)' },
+    { id: 'emitter-events', displayName: 'Events emitter (jest-metadata)' },
+    { id: 'environment', displayName: 'Test Environment (jest-metadata)' },
+    { id: 'metadata', displayName: 'Metadata (jest-metadata)' },
+    { id: 'reporter', displayName: 'Reporter (jest-metadata)' },
+  );
+}
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const optimizeTracing: <F>(f: F) => F = isTraceEnabled() ? (f) => f : ((() => noop) as any);
